move: SearchInput을 search 폴더로 이동
